refactor(client): update project structure
