+++ client/.eslintrc.js
@@ -19,11 +19,6 @@ module.exports = {
   rules: {
     'prettier/prettier': ['error', { endOfLine: 'lf' }],
     'linebreak-style': ['error', 'unix'],
-    'vue/multi-word-component-names': [
-      'error',
-      {
-        ignores: ['Header', 'Tooltip'],
-      },
-    ],
+    'vue/multi-word-component-names': 'off',
   },
 };
