+++ client/src/components/generic-button/generic-button.vue
@@ -9,7 +9,7 @@ defineProps<{
 </template>
 
 <style scoped lang="scss">
-@import '../mediaqueries.scss';
+@import '../../mediaqueries.scss';
 
 button {
   font-family: unset;
