+++ client/src/components/loading-animation/loading-animation.vue
@@ -17,7 +17,7 @@
 </template>
 
 <style scoped lang="scss">
-@import '../mediaqueries.scss';
+@import '../../mediaqueries.scss';
 
 .loading div {
   position: absolute;
