+++ client/src/components/player-info/player-info.vue
@@ -17,7 +17,7 @@ defineProps<{
 </template>
 
 <style scoped lang="scss">
-@import '../mediaqueries.scss';
+@import '../../mediaqueries.scss';
 
 .player-info {
   display: flex;
