+++ client/src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
-import App from './App.vue';
+import App from './app.vue';
 
 const pinia = createPinia();
 const app = createApp(App);
