+++ client/vite.config.ts
@@ -2,7 +2,6 @@
 import { defineConfig } from 'vite';
 import vue from '@vitejs/plugin-vue';
 import replace from '@rollup/plugin-replace';
-
 // https://vitejs.dev/config/
 export default defineConfig({
   optimizeDeps: {
@@ -21,7 +20,7 @@ export default defineConfig({
     dedupe: ['vue'],
   },
   test: {
-    include: ['./tests/**/*.test.ts'],
+    include: ['./**/*.test.ts'],
     globals: true,
     environment: 'happy-dom',
     silent: true,
@@ -56,3 +55,6 @@ export default defineConfig({
     },
   },
 });
+function fileURLToPath(arg0: URL): string {
+  throw new Error('Function not implemented.');
+}
